refactor(vscode): split webview messages.ts into per-domain files
Splits the 2755-line webview-ui/src/types/messages.ts into 14 domain files
under types/messages/ with a barrel index. Type definitions are preserved
byte-identically; consumer imports are unchanged thanks to the barrel.
