+++ packages/kilo-vscode/webview-ui/src/types/messages/index.ts
@@ -1,0 +1,17 @@
+/**
+ * Types for extension <-> webview message communication
+ */
+
+export * from "./connection"
+export * from "./parts"
+export * from "./sessions"
+export * from "./permissions"
+export * from "./questions"
+export * from "./providers"
+export * from "./agents"
+export * from "./config"
+export * from "./profile"
+export * from "./agent-manager"
+export * from "./migration"
+export * from "./extension-messages"
+export * from "./webview-messages"
