+++ packages/kilo-vscode/webview-ui/src/types/messages/parts.ts
@@ -1,0 +1,94 @@
+// Tool state for tool parts
+export type ToolState =
+  | { status: "pending"; input: Record<string, unknown> }
+  | { status: "running"; input: Record<string, unknown>; title?: string }
+  | { status: "completed"; input: Record<string, unknown>; output: string; title: string }
+  | { status: "error"; input: Record<string, unknown>; error: string }
+
+// Base part interface - all parts have these fields
+export interface BasePart {
+  id: string
+  sessionID?: string
+  messageID?: string
+}
+
+// Part types from the backend
+export interface TextPart extends BasePart {
+  type: "text"
+  text: string
+}
+
+export interface FilePartSource {
+  type: "file"
+  path: string
+  text: {
+    value: string
+    start: number
+    end: number
+  }
+}
+
+export interface FilePart extends BasePart {
+  type: "file"
+  mime: string
+  url: string
+  filename?: string
+  source?: FilePartSource
+}
+
+export interface ToolPart extends BasePart {
+  type: "tool"
+  tool: string
+  state: ToolState
+}
+
+export interface ReasoningPart extends BasePart {
+  type: "reasoning"
+  text: string
+}
+
+// Step parts from the backend
+export interface StepStartPart extends BasePart {
+  type: "step-start"
+}
+
+export interface StepFinishPart extends BasePart {
+  type: "step-finish"
+  reason?: string
+  cost?: number
+  tokens?: {
+    input: number
+    output: number
+    reasoning?: number
+    cache?: { read: number; write: number }
+  }
+}
+
+export type Part = TextPart | FilePart | ToolPart | ReasoningPart | StepStartPart | StepFinishPart
+
+// Part delta for streaming updates
+export interface PartDelta {
+  type: "text-delta"
+  textDelta?: string
+}
+
+// Token usage for assistant messages
+export interface TokenUsage {
+  input: number
+  output: number
+  reasoning?: number
+  cache?: { read: number; write: number }
+}
+
+// Context usage derived from the last assistant message's tokens
+export interface ContextUsage {
+  tokens: number
+  percentage: number | null
+}
+
+export interface FileAttachment {
+  mime: string
+  url: string
+  filename?: string
+  source?: FilePartSource
+}
