+++ packages/kilo-vscode/webview-ui/src/types/messages/profile.ts
@@ -1,0 +1,29 @@
+// Kilo notification types (mirrored from kilo-gateway)
+export interface KilocodeNotificationAction {
+  actionText: string
+  actionURL: string
+}
+
+export interface KilocodeNotification {
+  id: string
+  title: string
+  message: string
+  action?: KilocodeNotificationAction
+  showIn?: string[]
+  suggestModelId?: string
+}
+
+// Profile types from kilo-gateway
+export interface KilocodeBalance {
+  balance: number
+}
+
+export interface ProfileData {
+  profile: {
+    email: string
+    name?: string
+    organizations?: Array<{ id: string; name: string; role: string }>
+  }
+  balance: KilocodeBalance | null
+  currentOrgId: string | null
+}
